mainに機能を移動させた｡
